feat: move to individual lodash packages
Moves to using individual lodash imports to reduce package size.
