+++ packages/inquirer/package.json
@@ -64,7 +64,12 @@
     "chalk": "^5.3.0",
     "cli-width": "^4.1.0",
     "external-editor": "^3.1.0",
-    "lodash": "^4.17.21",
+    "lodash.clone": "^4.5.0",
+    "lodash.defaults": "^4.2.0",
+    "lodash.filter": "^4.6.0",
+    "lodash.get": "^4.4.2",
+    "lodash.isplainobject": "^4.0.6",
+    "lodash.set": "^4.3.2",
     "mute-stream": "1.0.0",
     "ora": "^5.4.1",
     "run-async": "^3.0.0",
